Speed up processing of dump files but adding caching layers into loading and using symbols.
The main bottleneck is the calls into the dbghlp SymXxxx functions which are slow which calling for the same symbol information over and over again.  So fix this by adding layers of symbol caching layers so that it tries to only call the symbol information once only for the same symbol.

Fix heap_entry::get_unused_bytes() returning negative values.
Add -out option to MiniDumper to generate output file instead of to the console