Break up more DbgHelpUtils into overridable components with only the in-build strings.
Standardise on calling the overridable source files *_resource.cpp so that they can be easy to spot.
Attempt to minimize the code in *_resource.cpp files to to reuse as much as possible the library (i.e. be mainly string tables or string references only - no logic).
Add DbgHelpUtils override resources to MiniDumpExplorer using the app resource lookup.