Most recent versions of DbgHelp have reinforced checks
Quite a bit of time was lost trying to understand why the code was not working on a Windows11 machine.
As it turns, newer Windows SDKs (and Win11), have changed the way DbgHelp parses the headers.
After reverse-engineering the code a bit, I finally found out that what changed is that `SizeOfData` is now actually checked. Not only is it checked, but it does so by also making sure that each debug directory entry not only is properly aligned... but its size too.
This means that for each `sizeof(IMAGE_DEBUG_DIRECTORY) + dir.SizeOfData` (it is an array) needs to be of a size aligned to 8 bytes. For future reference, the culprit is `DbgHelp.dll!ReadCallerData`.
I think it's probably a mistake, but this is easily fixed on our side once we know about it. (Who'd have guessed, you had a 1/8 chance of having it working depending on the pdb path length!)